상품 상세 페이지 작업  - 상품 상세 페이지 추가  - 상품 목록에서 상세 페이지로 넘어갈 수 있도록 개선  - 주문하기 로직 수정

+++ src/main/java/fromryw/lpshop/order/service/BaseOrderService.java
@@ -103,6 +103,8 @@ public class BaseOrderService implements OrderService {
         orderItemService.saveAll(newOrderItems);
 
         // 장바구니 데이터 삭제(특정 회원)
-        cartService.removeAll(order.getMemberId());
+        if ("cart".equals(orderReq.getOrderType())) {
+            cartService.removeAll(order.getMemberId());
+        }
     }
 }
